+++ moyo/src/symmetrize.rs
@@ -1,9 +1,10 @@
+mod conventional_cell;
 mod layer_standardize;
 mod magnetic_standardize;
 mod standardize;
+mod wyckoff;
 
 pub(super) use layer_standardize::StandardizedLayerCell;
 pub(super) use magnetic_standardize::StandardizedMagneticCell;
-pub(super) use standardize::{
-    StandardizedCell, orbits_in_cell, wyckoff_positions_under_normalizer,
-};
+pub(super) use standardize::StandardizedCell;
+pub(super) use wyckoff::{orbits_in_cell, wyckoff_positions_under_normalizer};
