+++ moyo/src/symmetrize/standardize.rs
@@ -1,21 +1,21 @@
-use itertools::{Itertools, iproduct};
-use log::{debug, warn};
+use log::warn;
 use nalgebra::linalg::{Cholesky, QR};
 use nalgebra::{Matrix3, Vector3, vector};
-use once_cell::sync::Lazy;
-use std::cmp::Ordering;
 use std::collections::HashMap;
 
+use super::conventional_cell::{
+    standardize_monoclinic_conv_cell, standardize_orthorhombic_conv_cell,
+};
+use super::wyckoff::{assign_wyckoffs_by_orbit, group_sites_by_orbit, match_wyckoff_coordinates};
 use crate::base::{
     Cell, EPS, Lattice, Linear, MoyoError, Operations, Permutation, Position, Rotations,
-    Transformation, UnimodularTransformation, orbits_from_permutations, project_rotations,
+    Transformation, UnimodularTransformation, project_rotations,
 };
 use crate::data::{
-    Centering, HallNumber, HallSymbol, LatticeSystem, WyckoffPosition, WyckoffPositionSpace,
-    arithmetic_crystal_class_entry, hall_symbol_entry, iter_wyckoff_positions,
+    HallNumber, HallSymbol, LatticeSystem, WyckoffPosition, arithmetic_crystal_class_entry,
+    hall_symbol_entry, iter_wyckoff_positions,
 };
 use crate::identify::SpaceGroup;
-use crate::math::SNF;
 
 pub struct StandardizedCell {
     // ------------------------------------------------------------------------
@@ -239,75 +239,6 @@ impl StandardizedCell {
     }
 }
 
-/// * `prim_num_atoms` - Number of atoms in the primitive cell
-/// * `prim_permutations` - Permutations of the primitive cell
-/// * `site_mapping` - Mapping from the site in cell to that in its primitive cell
-pub fn orbits_in_cell(
-    prim_num_atoms: usize,
-    prim_permutations: &[Permutation],
-    site_mapping: &[usize],
-) -> Vec<usize> {
-    // prim_orbits: [prim_num_atoms] -> [prim_num_atoms]
-    let prim_orbits = orbits_from_permutations(prim_num_atoms, prim_permutations);
-
-    let num_atoms = site_mapping.len();
-    let mut map = HashMap::new();
-    let mut orbits = vec![]; // [num_atoms] -> [num_atoms]
-    for i in 0..num_atoms {
-        // site_mapping: [num_atoms] -> [prim_num_atoms]
-        let key = prim_orbits[site_mapping[i]]; // in [prim_num_atoms]
-        map.entry(key).or_insert(i);
-        orbits.push(*map.get(&key).unwrap());
-    }
-    orbits
-}
-
-/// Group sites of a conventional cell by crystallographic orbit. Shared
-/// between the bulk and layer pipelines.
-pub(super) struct OrbitGrouping {
-    /// `[std_num_atoms] -> [num_orbits]`: orbit index for each site.
-    pub mapping: Vec<usize>,
-    /// `[num_orbits] -> [std_num_atoms]`: representative site for each orbit
-    /// (the lowest-index site in the orbit). Used only for diagnostics.
-    pub remapping: Vec<usize>,
-    /// Number of sites per orbit, indexed by orbit. `multiplicities.len()` is
-    /// the orbit count.
-    pub multiplicities: Vec<usize>,
-}
-
-pub(super) fn group_sites_by_orbit(
-    prim_num_atoms: usize,
-    prim_permutations: &[Permutation],
-    site_mapping: &[usize],
-    std_num_atoms: usize,
-) -> OrbitGrouping {
-    let orbits = orbits_in_cell(prim_num_atoms, prim_permutations, site_mapping);
-
-    let mut num_orbits = 0;
-    let mut mapping = vec![0; std_num_atoms];
-    let mut remapping = vec![];
-    for i in 0..std_num_atoms {
-        if orbits[i] == i {
-            mapping[i] = num_orbits;
-            remapping.push(i);
-            num_orbits += 1;
-        } else {
-            mapping[i] = mapping[orbits[i]];
-        }
-    }
-
-    let mut multiplicities = vec![0; num_orbits];
-    for i in 0..std_num_atoms {
-        multiplicities[mapping[i]] += 1;
-    }
-
-    OrbitGrouping {
-        mapping,
-        remapping,
-        multiplicities,
-    }
-}
-
 /// Align a set of primitive-cell permutations (as produced by the symmetry
 /// search) with a target operation list (as produced by traversing a Hall
 /// symbol) by matching rotation matrices. Both pipelines need this because
@@ -339,117 +270,6 @@ pub(super) fn align_primitive_permutations(
         .collect()
 }
 
-/// Run the per-orbit Wyckoff-assignment loop over a precomputed
-/// `OrbitGrouping`. The closure produces the database-specific Wyckoff
-/// representative for an `(orbit_position, multiplicity)` pair (returning
-/// `None` if no orbit in the database matches that multiplicity at that
-/// position). Shared between bulk and layer pipelines: the only thing that
-/// differs is the database lookup, which the caller passes in as the
-/// closure.
-pub(super) fn assign_wyckoffs_by_orbit<W, F>(
-    group: &OrbitGrouping,
-    positions: &[Position],
-    mut find_for_orbit: F,
-) -> Result<Vec<W>, MoyoError>
-where
-    W: Clone,
-    F: FnMut(&Position, usize) -> Option<W>,
-{
-    let mut representative_wyckoffs: Vec<Option<W>> = vec![None; group.multiplicities.len()];
-    for (i, position) in positions.iter().enumerate() {
-        let orbit = group.mapping[i];
-        if representative_wyckoffs[orbit].is_some() {
-            continue;
-        }
-        if let Some(w) = find_for_orbit(position, group.multiplicities[orbit]) {
-            representative_wyckoffs[orbit] = Some(w);
-        }
-    }
-
-    for (orbit, wyckoff) in representative_wyckoffs.iter().enumerate() {
-        if wyckoff.is_none() {
-            debug!(
-                "Failed to assign Wyckoff position with multiplicity {} at representative site {}",
-                group.multiplicities[orbit], group.remapping[orbit]
-            );
-        }
-    }
-    let representative_wyckoffs = representative_wyckoffs
-        .into_iter()
-        .map(|w| w.ok_or(MoyoError::WyckoffPositionAssignmentError))
-        .collect::<Result<Vec<_>, _>>()?;
-
-    Ok(group
-        .mapping
-        .iter()
-        .map(|&orbit| representative_wyckoffs[orbit].clone())
-        .collect())
-}
-
-/// For each conventional-basis normalizer operation, apply it to `positions`
-/// and recompute the Wyckoff positions of the structure.
-///
-/// Returns one inner `Vec` per operation (same order as `conventional_ops`),
-/// each holding one [`WyckoffPosition`] per atom in `positions` order. The
-/// result is RAW: duplicate assignments produced by different operations are
-/// NOT removed (the caller deduplicates / canonicalizes).
-///
-/// Contract: `conventional_ops`, `lattice`, and `positions` must all be in the
-/// conventional standardized setting (the Wyckoff database coordinates are);
-/// `site_orbits[i]` is the orbit index of atom `i` and `orbit_multiplicities`
-/// is indexed by orbit. The orbit structure is invariant under the normalizer,
-/// so it is computed once by the caller and reused for every operation. The
-/// lattice is normalizer-invariant (`P^T M P = M`), so it is used unchanged for
-/// the Cartesian-distance tolerance inside [`match_wyckoff_coordinates`].
-pub(crate) fn wyckoff_positions_under_normalizer(
-    conventional_ops: &[UnimodularTransformation],
-    lattice: &Lattice,
-    positions: &[Position],
-    hall_number: HallNumber,
-    site_orbits: &[usize],
-    orbit_multiplicities: &[usize],
-    symprec: f64,
-) -> Result<Vec<Vec<WyckoffPosition>>, MoyoError> {
-    let num_orbits = orbit_multiplicities.len();
-
-    let mut result = Vec::with_capacity(conventional_ops.len());
-    for op in conventional_ops {
-        let linear = op.linear.map(|e| e as f64);
-
-        // Apply the active map `x -> P x + p` (reduced mod 1) to each atom and
-        // assign a Wyckoff position per orbit. `match_wyckoff_coordinates` only
-        // matches the database representative coordinate (up to lattice
-        // translations and the free-variable span), so -- as in
-        // `assign_wyckoffs_by_orbit` -- each orbit is tried against successive
-        // atoms until one lands on that representative.
-        let mut representative_wyckoffs: Vec<Option<WyckoffPosition>> = vec![None; num_orbits];
-        for (i, &orbit) in site_orbits.iter().enumerate() {
-            if representative_wyckoffs[orbit].is_some() {
-                continue;
-            }
-            let transformed = (linear * positions[i] + op.origin_shift).map(|e| e.rem_euclid(1.0));
-            if let Some(wyckoff) = iter_wyckoff_positions(hall_number, orbit_multiplicities[orbit])
-                .find(|w| match_wyckoff_coordinates(&transformed, w.coordinates, lattice, symprec))
-                .cloned()
-            {
-                representative_wyckoffs[orbit] = Some(wyckoff);
-            }
-        }
-
-        let representative_wyckoffs = representative_wyckoffs
-            .into_iter()
-            .map(|w| w.ok_or(MoyoError::WyckoffPositionAssignmentError))
-            .collect::<Result<Vec<_>, _>>()?;
-        let assignment = site_orbits
-            .iter()
-            .map(|&orbit| representative_wyckoffs[orbit].clone())
-            .collect();
-        result.push(assignment);
-    }
-
-    Ok(result)
-}
-
 /// Niggli reduction for distorted triclinic lattice systems is numerically so challenging.
 /// Thus, we skip checking reduction condition.
 fn standardize_triclinic_cell(
@@ -462,270 +282,6 @@ fn standardize_triclinic_cell(
         niggli_linear * transformation_to_prim_std.linear,
         transformation_to_prim_std.origin_shift,
     )
-}
-
-/// Candidate unimodular corrections with entries in `[-1, 1]`.
-///
-/// This bounded search is used as a best-effort search space for monoclinic
-/// conventional-cell standardization. Exhaustiveness is not claimed here.
-static UNIMODULAR3_RANGE1: Lazy<Vec<UnimodularTransformation>> = Lazy::new(|| {
-    (0..9)
-        .map(|_| -1_i32..=1_i32)
-        .multi_cartesian_product()
-        .filter_map(|v| {
-            let mat = Matrix3::new(v[0], v[1], v[2], v[3], v[4], v[5], v[6], v[7], v[8]);
-            let det = mat.map(|e| e as f64).determinant().round() as i32;
-            if det == 1 {
-                Some(UnimodularTransformation::from_linear(mat))
-            } else {
-                None
-            }
-        })
-        .collect()
-});
-
-/// The six axis permutations as proper (det = +1) signed-permutation matrices. The
-/// three odd permutations (transpositions) are made right-handed by negating their
-/// first axis, since [`Transformation`] requires a positive determinant.
-///
-/// One proper representative per permutation is sufficient: for every orthorhombic
-/// space group, whether a permutation preserves the operation set and centering does
-/// not depend on the choice of axis signs (the point-group rotations are diagonal, so
-/// conjugation by a sign flip leaves them unchanged, and centering half-translations
-/// absorb the sign of any non-symmorphic translation). The sign-flipped axes are
-/// re-oriented away later by `symmetrize_lattice`.
-static AXIS_PERMUTATIONS3: Lazy<Vec<UnimodularTransformation>> = Lazy::new(|| {
-    [
-        // Even permutations (already det = +1).
-        Matrix3::new(1, 0, 0, 0, 1, 0, 0, 0, 1), // abc (identity)
-        Matrix3::new(0, 1, 0, 0, 0, 1, 1, 0, 0), // cab
-        Matrix3::new(0, 0, 1, 1, 0, 0, 0, 1, 0), // bca
-        // Odd permutations, first axis negated to restore det = +1.
-        Matrix3::new(0, 1, 0, -1, 0, 0, 0, 0, 1), // b(-a)c  (a <-> b)
-        Matrix3::new(-1, 0, 0, 0, 0, 1, 0, 1, 0), // (-a)cb  (b <-> c)
-        Matrix3::new(0, 0, 1, 0, 1, 0, -1, 0, 0), // c b(-a) (a <-> c)
-    ]
-    .into_iter()
-    .map(UnimodularTransformation::from_linear)
-    .collect()
-});
-
-/// Standardize monoclinic cell by choosing reduced basis vectors for perpendicular plane to the unique axis while keeping matrix representations of `conv_std_generators`.
-fn standardize_monoclinic_conv_cell(
-    prim_lattice: &Lattice,
-    transformation_to_prim_std: &UnimodularTransformation,
-    centering: Centering,
-    conv_std_generators: &Operations,
-    epsilon: f64,
-) -> Linear {
-    let mut candidate_conv_transformations = vec![];
-    // Search candidate corrections in the bounded `[-1, 1]` window and choose
-    // the least skewed valid one. This is a best-effort bounded search.
-    for trans_corr in UNIMODULAR3_RANGE1.iter() {
-        // trans_corr should keep centering translations
-        if centering.lattice_points().iter().any(|translation| {
-            let mut diff = trans_corr.linear.map(|e| e as f64) * translation - translation;
-            diff -= diff.map(|e| e.round()); // in [-0.5, 0.5]
-            diff.iter().any(|e| e.abs() > epsilon)
-        }) {
-            continue;
-        }
-
-        // trans_corr should keep the matrix representations of `conv_std_generators`.
-        if conv_std_generators.iter().any(|ops| {
-            let corr_ops = trans_corr.transform_operation(ops);
-            if corr_ops.rotation != ops.rotation {
-                true
-            } else {
-                let mut diff = corr_ops.translation - ops.translation;
-                diff -= diff.map(|e| e.round()); // in [-0.5, 0.5]
-                diff.iter().any(|e| e.abs() > epsilon)
-            }
-        }) {
-            continue;
-        }
-
-        let refined_conv_trans = Transformation::new(
-            transformation_to_prim_std.linear * centering.linear() * trans_corr.linear,
-            transformation_to_prim_std.origin_shift,
-        );
-        let refined_conv_lattice = refined_conv_trans.transform_lattice(prim_lattice);
-        let cos_angles = refined_conv_lattice.lattice_constant()[3..]
-            .iter()
-            .map(|angle_deg| angle_deg.to_radians().cos())
-            .collect::<Vec<_>>();
-        // `skewness` gets smaller as the basis vectors are closer to orthorhombic.
-        let skewness = cos_angles.iter().map(|cos| cos.abs()).sum::<f64>();
-        // The monoclinic angle and its supplement give the same `skewness`
-        // (`|cos|` is symmetric about 90 deg), so the unimodular correction that
-        // flips the angle to its supplement is an equally valid candidate. Among
-        // such ties we follow the ITA convention and prefer a non-acute angle,
-        // i.e. the smaller (more negative) signed cosine sum.
-        let signed_cos_sum = cos_angles.iter().sum::<f64>();
-        candidate_conv_transformations.push((
-            skewness,
-            signed_cos_sum,
-            centering.linear() * trans_corr.linear,
-        ));
-    }
-
-    candidate_conv_transformations
-        .into_iter()
-        .min_by(|(skewness_lhs, cos_lhs, _), (skewness_rhs, cos_rhs, _)| {
-            if (skewness_lhs - skewness_rhs).abs() < EPS {
-                cos_lhs.partial_cmp(cos_rhs).unwrap()
-            } else {
-                skewness_lhs.partial_cmp(skewness_rhs).unwrap()
-            }
-        })
-        .unwrap()
-        .2
-}
-
-/// Standardize an orthorhombic conventional cell by choosing the axis permutation
-/// that orders the basis-vector lengths as `a <= b <= c` as much as possible, while
-/// keeping the matrix representations of the symmetry operations (as a set) and the
-/// centering.
-///
-/// Unlike the monoclinic case, an axis permutation relabels generators among
-/// themselves, so the *set* of `conv_std_operations` -- not each individual generator
-/// -- must be preserved. Permutations that would change the operation set or move a
-/// centered face (e.g. for `C`-centering) are rejected, which reproduces seekpath's
-/// "number of distinct projections" without a hardcoded per-space-group table.
-fn standardize_orthorhombic_conv_cell(
-    prim_lattice: &Lattice,
-    transformation_to_prim_std: &UnimodularTransformation,
-    centering: Centering,
-    conv_std_operations: &Operations,
-    epsilon: f64,
-) -> Linear {
-    let mut candidate_conv_transformations = vec![];
-    // Consider the six axis permutations and keep those that preserve the centering
-    // and the operation set; pick the one giving the smallest lengths.
-    for trans_perm in AXIS_PERMUTATIONS3.iter() {
-        // trans_perm should keep centering translations.
-        if centering.lattice_points().iter().any(|translation| {
-            let mut diff = trans_perm.linear.map(|e| e as f64) * translation - translation;
-            diff -= diff.map(|e| e.round()); // in [-0.5, 0.5]
-            diff.iter().any(|e| e.abs() > epsilon)
-        }) {
-            continue;
-        }
-
-        // trans_perm should keep the *set* of conv_std_operations: each conjugated
-        // operation must coincide with some operation in the set (rotation exactly,
-        // translation modulo the centering lattice). `conv_std_operations` holds one
-        // coset representative per rotation, so translations are compared modulo the
-        // centering translations, not only modulo 1.
-        if conv_std_operations.iter().any(|ops| {
-            let perm_ops = trans_perm.transform_operation(ops);
-            !conv_std_operations.iter().any(|other| {
-                perm_ops.rotation == other.rotation
-                    && translations_equal_mod_centering(
-                        &perm_ops.translation,
-                        &other.translation,
-                        centering,
-                        epsilon,
-                    )
-            })
-        }) {
-            continue;
-        }
-
-        let refined_conv_trans = Transformation::new(
-            transformation_to_prim_std.linear * centering.linear() * trans_perm.linear,
-            transformation_to_prim_std.origin_shift,
-        );
-        let lc = refined_conv_trans
-            .transform_lattice(prim_lattice)
-            .lattice_constant();
-        let lengths = [lc[0], lc[1], lc[2]];
-        candidate_conv_transformations.push((lengths, centering.linear() * trans_perm.linear));
-    }
-
-    // Choose the lexicographically smallest (a, b, c). `min_by` keeps the first
-    // element among EPS-ties, so iteration order makes the selection stable.
-    candidate_conv_transformations
-        .into_iter()
-        .min_by(|(lengths_lhs, _), (lengths_rhs, _)| {
-            lengths_lhs
-                .iter()
-                .zip(lengths_rhs.iter())
-                .find_map(|(lhs, rhs)| {
-                    if (lhs - rhs).abs() < EPS {
-                        None
-                    } else {
-                        Some(lhs.partial_cmp(rhs).unwrap())
-                    }
-                })
-                .unwrap_or(Ordering::Equal)
-        })
-        .unwrap()
-        .1
-}
-
-/// Whether two fractional translations are equal modulo the centering lattice (which
-/// includes the integer lattice). Used to compare symmetry operations of a
-/// conventional centered cell, where a given coset may be represented by translations
-/// differing by a centering vector.
-fn translations_equal_mod_centering(
-    lhs: &Vector3<f64>,
-    rhs: &Vector3<f64>,
-    centering: Centering,
-    epsilon: f64,
-) -> bool {
-    let mut diff = lhs - rhs;
-    diff -= diff.map(|e| e.round()); // in [-0.5, 0.5]
-    centering.lattice_points().iter().any(|lattice_point| {
-        let mut residual = diff - lattice_point;
-        residual -= residual.map(|e| e.round());
-        residual.iter().all(|e| e.abs() <= epsilon)
-    })
-}
-
-/// Test whether `position` matches a Wyckoff orbit described by `coordinates`
-/// (the parsed-coordinates string from a Wyckoff entry, e.g. `"x,1/2,z"`)
-/// modulo a bounded integer offset. Shared between the bulk and layer
-/// pipelines: both Wyckoff databases store their representative coordinate as
-/// the same `&str` shape, so the SNF-based offset search is identical.
-///
-/// The search probes offsets in `[-1, 1]^3` first, then the remaining shell
-/// in `[-2, 2]^3`. The math (suppressing tolerances): we want `y, offset` so
-/// that `lattice * (space.linear * y + space.origin - position - offset)` is
-/// near zero. Decomposing `space.linear` as `D = L * space.linear * R` (SNF)
-/// reduces the integer search to picking `D^{-1} L (offset + position -
-/// space.origin)` and reading off whether the residual is small in Cartesian
-/// distance.
-pub(super) fn match_wyckoff_coordinates(
-    position: &Position,
-    coordinates: &str,
-    lattice: &Lattice,
-    symprec: f64,
-) -> bool {
-    let space = WyckoffPositionSpace::new(coordinates);
-    let snf = SNF::new(&space.linear);
-
-    let iter_multi_1 = iproduct!(-1..=1, -1..=1, -1..=1);
-    let iter_multi_2 = iproduct!(-2_i32..=2_i32, -2_i32..=2_i32, -2_i32..=2_i32)
-        .filter(|&(n1, n2, n3)| n1.abs() == 2 || n2.abs() == 2 || n3.abs() == 2);
-
-    for offset in iter_multi_1.chain(iter_multi_2) {
-        let offset = Vector3::new(offset.0 as f64, offset.1 as f64, offset.2 as f64);
-        let b = snf.l.map(|e| e as f64) * (offset + position - space.origin);
-        let mut rinvy = Vector3::zeros();
-        for i in 0..3 {
-            if snf.d[(i, i)] != 0 {
-                rinvy[i] = b[i] / snf.d[(i, i)] as f64;
-            }
-        }
-
-        let y = snf.r.map(|e| e as f64) * rinvy;
-        let diff = space.linear.map(|e| e as f64) * y + space.origin - position - offset;
-        if lattice.cartesian_coords(&diff).norm() < symprec {
-            return true;
-        }
-    }
-    false
 }
 
 /// Symmetrize positions by site symmetry groups. Operates on a slice rather
